Makes dynamic the obstacle layout instead of being hard-coded in the SVG
In attempting to explore the curve layout algorithm I made several, what I
thought were, simplifying assumptions. Working on the algorithm it
turned out that these just got in the way, rather than made things
easier. One of them was that I would work with the SVG way of y=0 at the
top of the frame, even though this wasn't the way I wanted to work with
things in the long term. That turned out to just make angles more
confusing than they needed to be so I'm going back on that
simplification. The simplest use case for dealing with the world ->
SVG transform is with static objects so I'm migrating the obstacles to
be templated values rather than hard-coded.