Begins test-driving out the course path calculation
The initial incarnation of the path calculations was thrown together and
tested by viewing the output rendered course. It's now time to go back
to the beginning and test drive the calculations properly.

This commit adds the first couple of tests. It only covers cases  where
boundary circles of both obstacles have identical radii.