Rephrases the path calculations in terms of the obstacle boundary circles
Not all obstacles (in the near future) are going to have their boundary
circles centred on the origin of the obstacle. This will be especially
true for those that have entry/exit of EITHER and so will have 2 or more
boundary circles definitely not centred on the origin.

I _think_ the best way to deal with this is to separate in the path
calculator the idea that the obstacle has origin/radius/entry/exit from
the idea that the boundary circle is also an 'obstacle' in that it also
has origin/radius/entry/exit. If early in the calculation the boundary
circle/s are found/defined as first class citizens then the calculation
can proceed as before, and to a large extent ignore the originating
obstacles themselves. At that point, the original obstacles are a
rendering concern rather than a path calculation concern.