Separates two of the concerns in the obstacle constructor
The deserialisation of property strings (eg. origin, entry, exit) was
happening at the same time as the checking/adding defaults for these
properies. This meant that the earlier stage of creating named obstacles
could not access eg. origin as a vector in order to use it.

The obstacle constructor now:
* converts those types that need to be deserialized string -> other
first into their native types
* then applies the constructors for named obstacles
* and finally adds defaults, if not already present