Corrects some problems with the use of the obstacle orientations
Storing the obstacle orientation parameter as a vector seemed a sensible
thing at the time but it's proved fiddly to then use it effectively.
I've migrated to storing the orientation just as a bearing and using it
as a rotation factor. It's stored as degrees, rather than radians,
because:
* the obstacle constructor is meant to be idempotent - ie. passing the
output of a previous constructor call back to the constructor after
modifying some param/s should be a valid thing to do
* its more intuitive to use degrees on the initial constructor call
* if the initial param in degrees is converted to radians then the
constructor can no longer be idempotent (calling it a second time with
radians orientation has different meaning)
* so storing and using the orientation as degrees is just easier