Quick and dirty conversion to an Express app
Long term, this is going to be a website. But I don't want to need any
server-side rendering or storage. I want it to run wholly in the users'
browser. So in order to support 'saving' of courses I think a sensible
approach might be to encode course info as base64'd JSON and add that as
a query param to site URL. Users can 'save' their course by being given a URL
containing all the course info at the time. And when they return, on
site load it can take the query param and extract the course.

At the moment, its still an app that renders a template server-side, but
its a small step in the right direction to managing the course data
correctly.