Adds the ability to mark 'out of bounds'
This introduces the concept of obstacles that don't participate in the
path of the course. Out of bounds markers are the obvious example of
these types of obstacles but there may also be cases where an obstacle
is desired to indicate something to the reader without influencing the
rendered path.

Out of bounds markers themselves have been added as 75% scale
yellow-with-black-stripe cones. The smaller scale (I think) helps
separate them from the rest of the course and not occur too much of the
readers focus.