Adds the rest of the obstacle-like properties to <left|right><Entry|Exit>Boundary
At the moment when an obstacle is decided to be not an 'either' obstacle
then the boundary circle is presumed to be centered on the origin of the
obstacle and have entry/exit same as the obstacle. This precludes
obstacles that have left/right/entry/exit boundaries, but are not of
entry/exit 'either'.

In order to be able to change this It's important that each boundary
circle has its own data set and the calculator works from this data,
rather than making presumptions about the obstacle.