Re-works the way boundary circles are managed in the obstacles
Previously I'd made a change to enforce all obstacles have both
leftXBoundaryCircle and rightXBoundaryCircle so that single-boundary
(LEFT or RIGHT) and EITHER obstacles were easier to handle (prior it
required maintaining an invariant between the entry/exit property and
whether or not there were left|rightXBoundaryCircle's which was
difficult to enforce).

That change is starting to feel a little clunky. It makes
single/multi-boundry obstacles more consistent but means the path
calculator has to understand the different between multi-boundary
obstacles that have a fixed side and those (EITHER) where the side needs
to be determined.

I'm going to try something different - let the entry/exit boundaries
properties be a single property that either takes on the single boundary
circle or is an array of all possible boundary circles. The path
calculation can then be ignorant of EITHER obstacles and act only on
whether there is one or more choices of boundary circles.