First prototype of the SVG grid in React
Rough flexbox layout to define the course grid and (what will probably
be a) properties box on the right hand side. Rough prototype of a grid
that has minor/major lines, expands to fill the viewport and has major
tick marks in the left and bottom margin.

Next: prototype zooming and panning of this grid.